refactor: redirect action_api imports to api.app, remove shim
