--- src/quodeq/action_api.py
@@ -1,19 +1,0 @@
-"""Re-export shim — moved to quodeq.api.app."""
-from quodeq.api.app import (
-    RateLimitStore,
-    InMemoryRateLimitStore,
-    create_rate_limit_store,
-    create_app,
-    main,
-)
-
-__all__ = [
-    "RateLimitStore",
-    "InMemoryRateLimitStore",
-    "create_rate_limit_store",
-    "create_app",
-    "main",
-]
-
-if __name__ == "__main__":
-    main()
