+++ tests/test_action_api_health.py
@@ -1,4 +1,4 @@
-from quodeq.action_api import create_app
+from quodeq.api.app import create_app
 
 
 def test_action_api_health():
